@@ -1,4 +1,7 @@
-﻿namespace MiniCRMCore.Areas.Offers.Models
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace MiniCRMCore.Areas.Offers.Models
 {
 	public enum OfferFileType
 	{
@@ -9,5 +12,23 @@
 		Certificate,
 		TechPassport,
 		Card
+	}
+
+	public enum OfferSectionType
+    {
+        Undefined = 0,
+        NotSet = 1,
+
+		Description,
+		SystemType,
+		ShortDescription,
+		Case,
+		Recommendations,
+		SimilarCases,
+		RestDocuments,
+		CoveringLetter,
+		TechPassport,
+		Certificates,
+		BusinessCard
 	}
 }
